game state fixes
Working on fixing all the minor bugs from before. This makes the set up
of the states better, and fixes a few minor bugs like friendly ghosts on
level one and actually going to the win/lose screens. I am going to be
altering the way textures work soon so that our load time will be
drastically decreased and we will be able to change splash screens
better.